finished refactoring pug api code
